requirements.txt
# Core Machine Learning Frameworks
torch>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
TA-Lib>=0.4.28

# NLP and Transformers
transformers>=4.30.0
tokenizers>=0.13.0

# Sentiment Analysis (FinBERT)
sentencepiece>=0.1.99

# Topic Modeling
gensim>=4.3.0

# Hyperparameter Optimization
optuna>=3.3.0

# Financial Data APIs
oandapyV20>=0.7.2
finnhub-python>=2.4.18

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress Bars
tqdm>=4.65.0

# Statistical Analysis
scipy>=1.10.0
statsmodels>=0.14.0

# SHAP for Feature Importance
shap>=0.42.0

# Time Series Utilities
darts>=0.24.0

# Imbalanced Learning
imbalanced-learn>=0.11.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# API Requests
requests>=2.31.0
aiohttp>=3.8.0

# Date/Time Handling
pytz>=2023.3
python-dateutil>=2.8.2

# Database (optional for data storage)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Jupyter for experimentation
jupyter>=1.0.0
ipykernel>=6.25.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
